Nits and refactoring
