+++ mozci/console/application.py
@@ -4,12 +4,14 @@ import sys
 
 from cleo import Application
 
+from mozci.console.commands.check_backfills import CheckBackfillsCommand
 from mozci.console.commands.decision import DecisionCommand
 from mozci.console.commands.push import PushCommands
 
 
 def cli():
     application = Application()
+    application.add(CheckBackfillsCommand())
     application.add(PushCommands())
     application.add(DecisionCommand())
     application.run()
